Add basic WikidPad extension
This adds a menu item "Publish Config" to the context menu of the main text control. The current page needs to have a location attribute for the command to be enabled. When run all the pages with that location attribute are rolled up and published into a single file. There are also basic tests for 'enabled', 'os', and 'host' to determine if the page should be rolled up.